requirements.txt
flask
flask_sqlalchemy
pycryptodome
flask_cors
